add process indicator
This change adds a CLI process indicator to bootstrap, exec, and
orchestrate when the wait flag is used. This change addresses a
UX request to provide a visual clue that operations are online when
waiting for specific jobs to complete.

The new indicator class will provide a BSD like spinner, with a status
pipe so that the operator knows what is happening and when.

Signed-off-by: Kevin Carter <kecarter@redhat.com>